Fixed rotation in (for instance) iPhone pictures
Quickly fixed rotation in pictures in the media page which have the
rotation in the jpeg exif metadata. This issue happens for example with
pictures from iphone.
Still not fixed for the album page but hey. The orientations buttons
still works even if it is counter-intuitive in this case, but that's not
very important.